added AccumulatedPoint.Default unit test. in order to do this I have moved the AccumulatedPoint class to it's own files.
