+++ src/Core/Geometry/AccumulatedPoint.h
@@ -24,41 +24,30 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
-#include "UnitTest.h"
+#pragma once
 
-#include "Core/Geometry/AccumulatedPoint.h"
-#include "Core/Geometry/PointCloud.h"
+#include <Eigen/Core>
 
-// ----------------------------------------------------------------------------
-//
-// ----------------------------------------------------------------------------
-TEST(AccumulatedPoint, Default)
+namespace open3d {
+
+class PointCloud;
+
+class AccumulatedPoint
 {
-    int size = 100;
+public:
+    AccumulatedPoint();
 
-    Eigen::Vector3d pmin(0.0, 0.0, 0.0);
-    Eigen::Vector3d pmax(1000.0, 1000.0, 1000.0);
+public:
+    void AddPoint(const PointCloud &cloud, int index);
+    Eigen::Vector3d GetAveragePoint() const;
+    Eigen::Vector3d GetAverageNormal() const;
+    Eigen::Vector3d GetAverageColor() const;
 
-    Eigen::Vector3d nmin(0.0, 0.0, 0.0);
-    Eigen::Vector3d nmax(1.0, 1.0, 1.0);
+private:
+    int num_of_points_;
+    Eigen::Vector3d point_;
+    Eigen::Vector3d normal_;
+    Eigen::Vector3d color_;
+};
 
-    Eigen::Vector3d cmin(0.0, 0.0, 0.0);
-    Eigen::Vector3d cmax(255.0, 255.0, 255.0);
-
-    open3d::PointCloud pc;
-
-    pc.points_.resize(size);
-    pc.normals_.resize(size);
-    pc.colors_.resize(size);
-
-    unit_test::Rand(pc.points_, pmin, pmax, 0);
-    unit_test::Rand(pc.normals_, nmin, nmax, 0);
-    unit_test::Rand(pc.colors_, cmin, cmax, 0);
-
-    open3d::AccumulatedPoint accpoint;
-    accpoint.AddPoint(pc, 0);
-
-    unit_test::ExpectEQ(839.215686, 392.156862, 780.392156, accpoint.GetAveragePoint());
-    unit_test::ExpectEQ(0.692861, 0.323767, 0.644296, accpoint.GetAverageNormal());
-    unit_test::ExpectEQ(214.0, 100.0, 199.0, accpoint.GetAverageColor());
-}
+}    // namespace open3d
