+++ src/Core/Geometry/DownSample.cpp
@@ -24,6 +24,7 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
+#include "AccumulatedPoint.h"
 #include "PointCloud.h"
 #include "TriangleMesh.h"
 
@@ -33,60 +34,6 @@
 #include <Core/Utility/Console.h>
 
 namespace open3d{
-
-namespace {
-
-class AccumulatedPoint
-{
-public:
-    AccumulatedPoint() :
-            num_of_points_(0),
-            point_(0.0, 0.0, 0.0),
-            normal_(0.0, 0.0, 0.0),
-            color_(0.0, 0.0, 0.0)
-    {
-    }
-
-public:
-    void AddPoint(const PointCloud &cloud, int index)
-    {
-        point_ += cloud.points_[index];
-        if (cloud.HasNormals()) {
-            if (!std::isnan(cloud.normals_[index](0)) &&
-                    !std::isnan(cloud.normals_[index](1)) &&
-                    !std::isnan(cloud.normals_[index](2))) {
-                normal_ += cloud.normals_[index];
-            }
-        }
-        if (cloud.HasColors()) {
-            color_ += cloud.colors_[index];
-        }
-        num_of_points_++;
-    }
-
-    Eigen::Vector3d GetAveragePoint() const
-    {
-        return point_ / double(num_of_points_);
-    }
-
-    Eigen::Vector3d GetAverageNormal() const
-    {
-        return normal_.normalized();
-    }
-
-    Eigen::Vector3d GetAverageColor() const
-    {
-        return color_ / double(num_of_points_);
-    }
-
-private:
-    int num_of_points_;
-    Eigen::Vector3d point_;
-    Eigen::Vector3d normal_;
-    Eigen::Vector3d color_;
-};
-
-}    // unnamed namespace
 
 std::shared_ptr<PointCloud> SelectDownSample(const PointCloud &input,
         const std::vector<size_t> &indices)
